restore some sanity
